# docs/swagger.yaml
basePath: /api
definitions:
  auth.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple user management
  title: User management
  version: "1.0"
paths:
  /change-password:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Change user account password
      parameters:
      - description: Current password
        in: formData
        minLength: 8
        name: current_password
        required: true
        type: string
      - description: New password
        in: formData
        minLength: 8
        name: new_password
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
        "400":
          description: Invalid input data
        "401":
          description: Unauthorized
        "500":
          description: Server error while changing password
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login account with email and password
      parameters:
      - description: User email address
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Tokens'
        "400":
          description: invalid input data
        "500":
          description: Server error while creating account
      summary: Login account
      tags:
      - auth
  /logout:
    post:
      consumes:
      - application/json
      description: Logout user account by invalidating refresh token
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
        "400":
          description: Invalid token format
        "401":
          description: Invalid refresh token
      security:
      - BearerAuth: []
      summary: Logout account
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user account with email and password
      parameters:
      - description: User email address
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
        "400":
          description: invalid input data
        "500":
          description: Server error while creating account
      summary: Register new account
      tags:
      - auth
  /renew-token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: formData
        name: refresh_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Tokens'
        "400":
          description: Invalid input data
        "401":
          description: Invalid refresh token
        "500":
          description: Server error while creating tokens
      summary: Renew access token
      tags:
      - auth
schemes:
- http
- https
swagger: "2.0"